pyproject.toml
[project]
name = "predoc-coding-sample"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "numpy>=1.26",
  "pandas>=2.2",
  "pyarrow>=15.0",
  "pyyaml>=6.0",
  "scikit-learn>=1.4",
  "statsmodels>=0.14",
  "matplotlib>=3.8",
  "pytest>=8.0",
]
